--- src/main/java/com/lop/ErrorEvent/ErrorInitiator.java
@@ -1,5 +1,0 @@
-package com.lop.ErrorEvent;
-
-public interface ErrorInitiator {
-    public void addListener(ErrorListener toAdd);
-}
