+++ src/main/java/com/lop/error_event/ErrorInitiator.java
@@ -1,0 +1,5 @@
+package com.lop.error_event;
+
+public interface ErrorInitiator {
+    void addListener(ErrorListener toAdd);
+}
