+++ src/main/java/com/lop/error_event/ErrorListener.java
@@ -1,4 +1,4 @@
-package com.lop.ErrorEvent;
+package com.lop.error_event;
 
 public interface ErrorListener {
     void errorOccurred(String message);
