+++ src/main/java/com/lop/exception/DataBaseException.java
@@ -1,4 +1,4 @@
-package com.lop.Exception;
+package com.lop.exception;
 
 public class DataBaseException extends Exception{
     public DataBaseException(String message) {
