+++ src/main/java/com/lop/exception/PasswordIncorrectException.java
@@ -1,4 +1,4 @@
-package com.lop.Exception;
+package com.lop.exception;
 
 public class PasswordIncorrectException extends Exception {
     public PasswordIncorrectException(String message) {
