+++ src/main/java/com/lop/exception/UnknownUserNameException.java
@@ -1,4 +1,4 @@
-package com.lop.Exception;
+package com.lop.exception;
 
 public class UnknownUserNameException extends Exception {
     public UnknownUserNameException(String message) {
