+++ src/main/java/com/lop/model/Beans/Bloc.java
@@ -1,4 +1,4 @@
-package com.lop.Model.Beans;
+package com.lop.model.Beans;
 
 public class Bloc {
     private int idBloc;
