+++ src/main/java/com/lop/model/Beans/Equipements.java
@@ -1,4 +1,4 @@
-package com.lop.Model.Beans;
+package com.lop.model.Beans;
 
 public class Equipements {
     private int numEqt;
