+++ src/main/java/com/lop/model/Beans/Evenements.java
@@ -1,4 +1,4 @@
-package com.lop.Model.Beans;
+package com.lop.model.Beans;
 
 import java.util.List;
