+++ src/main/java/com/lop/model/dao/Factory.java
@@ -1,4 +1,4 @@
-package com.lop.Model.dao;
+package com.lop.model.dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
