+++ src/main/java/com/lop/model/dao/PrintEvent.java
@@ -1,4 +1,4 @@
-package com.lop.Model.dao;
+package com.lop.model.dao;
 
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
