+++ src/main/java/com/lop/view/ConnectionWindow.java
@@ -1,4 +1,4 @@
-package com.lop.View;
+package com.lop.view;
 
 
 /*
