+++ src/main/java/com/lop/view/ConnectionWindow.jfd
