+++ src/main/java/com/lop/view/picture/DarkPurple.theme.json
