+++ src/main/java/com/lop/view/src/source/MyTableModel.java
@@ -1,4 +1,4 @@
-package com.lop.View.src.source;
+package com.lop.view.src.source;
 
 import java.util.ArrayList;
 import java.util.Arrays;
